+++ docs/source/changelog.rst
@@ -14,6 +14,7 @@ Development version
   of threads per process is roughly the same. Old default was to use one
   process and only threads, i.e. ``proccesses=1``, ``threads_per_process=cores``.
 - fix bug (forgotten async def) in ``OARCluster._submit_job`` (:pr:`380`).
+- ``LSFCluster``: switch to ``use_stdin=True`` (:pr:`388`).
 
 0.7.0 / 2019-10-09
 ------------------